+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Copy file directory feature (copies directory path with trailing separator)
+- E2E tests for Copy Parent Hash feature
 
 ## [1.1.2] - 2026-03-13